Move the 'v2' code in place
