+++ crates/credentials/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod pex;
 pub mod vc;
-pub mod vcv2;
